Fix pipeline failures
Recent update in golang.org/x/tools (which is a nested dependency
of golang.org/x/lint that Docker Machine project is using) related to
"Go 1.12 support is dropped" have caused our pipelines to start failing,
because we're still on Go 1.12.x.

The good news is that golang.org/x/lint itself points an older version
of golang.org/x/tools. The only thing we need to do is to enable usage
of Go Modules when installing the linter.

Which is what this commit does.